fix(volumes): Use UUID to mount volumes instead of volume id
EBS Volume ID will change when we start a new machine from an image. Since the ID is only known after the machine boots. We have to wait for the machine to boot and then run a playbook to reconfigure the mount.

But UUID remains the same. so, we don't need to change anything in the new machine at the first boot.

Reference: https://aws.amazon.com/blogs/compute/how-to-mount-linux-volume-and-keep-mount-point-consistency/